refactor(prediction-error): Move tests from file to folder
