+++ skore/tests/unit/sklearn/plot/prediction_error/test_comparison_estimator.py
@@ -1,0 +1,191 @@
+import matplotlib as mpl
+import numpy as np
+import pytest
+from skore import ComparisonReport, EstimatorReport
+from skore.sklearn._plot import PredictionErrorDisplay
+from skore.sklearn._plot.metrics.prediction_error import RangeData
+
+
+def test_prediction_error_comparison_report_display_regression(pyplot, regression_data):
+    """Check the attributes and default plotting behaviour of the prediction error plot
+    with a comparison report."""
+    estimator, X_train, X_test, y_train, y_test = regression_data
+    report = ComparisonReport(
+        reports={
+            "estimator 1": EstimatorReport(
+                estimator,
+                X_train=X_train,
+                y_train=y_train,
+                X_test=X_test,
+                y_test=y_test,
+            ),
+            "estimator 2": EstimatorReport(
+                estimator,
+                X_train=X_train,
+                y_train=y_train,
+                X_test=X_test,
+                y_test=y_test,
+            ),
+        },
+    )
+    display = report.metrics.prediction_error()
+    assert isinstance(display, PredictionErrorDisplay)
+
+    # check the structure of the attributes
+    assert display.estimator_names == ["estimator 1", "estimator 2"]
+    assert isinstance(display.y_true, list)
+    assert isinstance(display.y_pred, list)
+    assert isinstance(display.residuals, list)
+    assert len(display.y_true) == len(display.y_pred) == len(display.residuals) == 2
+    assert display.data_source == "test"
+    assert isinstance(display.range_y_true, RangeData)
+    assert isinstance(display.range_y_pred, RangeData)
+    assert isinstance(display.range_residuals, RangeData)
+    for attr in ("y_true", "y_pred", "residuals"):
+        global_min, global_max = np.inf, -np.inf
+        for display_attr in getattr(display, attr):
+            global_min = min(global_min, np.min(display_attr))
+            global_max = max(global_max, np.max(display_attr))
+        assert getattr(display, f"range_{attr}").min == global_min
+        assert getattr(display, f"range_{attr}").max == global_max
+
+    display.plot()
+    assert isinstance(display.line_, mpl.lines.Line2D)
+    assert display.line_.get_label() == "Perfect predictions"
+    assert display.line_.get_color() == "black"
+
+    assert isinstance(display.scatter_, list)
+    for scatter in display.scatter_:
+        assert isinstance(scatter, mpl.collections.PathCollection)
+
+    assert isinstance(display.ax_, mpl.axes.Axes)
+    legend = display.ax_.get_legend()
+    assert legend.get_title().get_text() == "Prediction errors on $\\bf{test}$ set"
+    assert len(legend.get_texts()) == 3
+
+    assert display.ax_.get_xlabel() == "Predicted values"
+    assert display.ax_.get_ylabel() == "Residuals (actual - predicted)"
+
+    assert display.ax_.get_aspect() not in ("equal", 1.0)
+
+
+def test_prediction_error_comparison_report_display_regression_kind(
+    pyplot, regression_data
+):
+    """Check the attributes when switching to the "actual_vs_predicted" kind."""
+    estimator, X_train, X_test, y_train, y_test = regression_data
+    report = ComparisonReport(
+        reports={
+            "estimator 1": EstimatorReport(
+                estimator,
+                X_train=X_train,
+                y_train=y_train,
+                X_test=X_test,
+                y_test=y_test,
+            ),
+            "estimator 2": EstimatorReport(
+                estimator,
+                X_train=X_train,
+                y_train=y_train,
+                X_test=X_test,
+                y_test=y_test,
+            ),
+        },
+    )
+    display = report.metrics.prediction_error()
+    display.plot(kind="actual_vs_predicted")
+    assert isinstance(display, PredictionErrorDisplay)
+
+    # check the structure of the attributes
+    assert isinstance(display.y_true, list)
+    assert isinstance(display.y_pred, list)
+    assert len(display.y_true) == len(display.y_pred) == 2
+    assert display.data_source == "test"
+
+    assert isinstance(display.line_, mpl.lines.Line2D)
+    assert display.line_.get_label() == "Perfect predictions"
+    assert display.line_.get_color() == "black"
+
+    assert isinstance(display.scatter_, list)
+    for scatter in display.scatter_:
+        assert isinstance(scatter, mpl.collections.PathCollection)
+
+    assert isinstance(display.ax_, mpl.axes.Axes)
+    legend = display.ax_.get_legend()
+    assert legend.get_title().get_text() == "Prediction errors on $\\bf{test}$ set"
+    assert len(legend.get_texts()) == 3
+
+    assert display.ax_.get_xlabel() == "Predicted values"
+    assert display.ax_.get_ylabel() == "Actual values"
+
+    assert display.ax_.get_aspect() in ("equal", 1.0)
+
+
+def test_prediction_error_display_comparison_estimator_kwargs(pyplot, regression_data):
+    """Check that we can pass keyword arguments to the prediction error plot when
+    there is a comparison report."""
+    estimator, X_train, X_test, y_train, y_test = regression_data
+    report = ComparisonReport(
+        reports={
+            "estimator 1": EstimatorReport(
+                estimator,
+                X_train=X_train,
+                y_train=y_train,
+                X_test=X_test,
+                y_test=y_test,
+            ),
+            "estimator 2": EstimatorReport(
+                estimator,
+                X_train=X_train,
+                y_train=y_train,
+                X_test=X_test,
+                y_test=y_test,
+            ),
+        },
+    )
+    display = report.metrics.prediction_error()
+    display.plot(
+        data_points_kwargs=[{"color": "red"}, {"color": "blue"}],
+        perfect_model_kwargs={"color": "orange"},
+    )
+    rgb_colors = [[[1.0, 0.0, 0.0, 0.3]], [[0.0, 0.0, 1.0, 0.3]]]
+    for scatter, rgb_color in zip(display.scatter_, rgb_colors):
+        np.testing.assert_allclose(scatter.get_facecolor(), rgb_color, rtol=1e-3)
+    assert display.line_.get_color() == "orange"
+
+
+@pytest.mark.parametrize("data_points_kwargs", ["not a list", [{"color": "red"}]])
+def test_prediction_error_comparison_estimator_kwargs_error(
+    pyplot, regression_data, data_points_kwargs
+):
+    """Check that we raise an error when we pass keyword arguments to the prediction
+    error plot if there is a comparison report."""
+    estimator, X_train, X_test, y_train, y_test = regression_data
+    report = ComparisonReport(
+        reports={
+            "estimator 1": EstimatorReport(
+                estimator,
+                X_train=X_train,
+                y_train=y_train,
+                X_test=X_test,
+                y_test=y_test,
+            ),
+            "estimator 2": EstimatorReport(
+                estimator,
+                X_train=X_train,
+                y_train=y_train,
+                X_test=X_test,
+                y_test=y_test,
+            ),
+        },
+    )
+    display = report.metrics.prediction_error()
+
+    err_msg = (
+        "You intend to plot prediction errors either from multiple estimators "
+        "or from a cross-validated estimator. We expect `data_points_kwargs` to be "
+        "a list of dictionaries with the same length as the number of "
+        "estimators or splits."
+    )
+    with pytest.raises(ValueError, match=err_msg):
+        display.plot(data_points_kwargs=data_points_kwargs)
